fix: uint16_t for coordinates inputs in ts_triggerarea

+++ Core/Inc/touchscreen.h
@@ -17,6 +17,6 @@
 
 
 void Touchscreen(void);
-static void ts_triggerArea(float x, float y, uint8_t state);
+static void ts_triggerArea(uint16_t x, uint16_t y, uint8_t state);
 
 #endif //MODAL_SYNTH_TOUCHSCREEN_H
